I can generate data now (1000 customers) to Couchbase And get them

+++ src/main/java/com/bachelor/boostr/model/BillingAccount.java
@@ -3,20 +3,16 @@ package com.bachelor.boostr.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 public class BillingAccount {
 
-    private BigDecimal balance;
+    private Double balance;
 
     private String passportData;
 
     private String contractNumber;
 
     private boolean isInDebt;
-
-
 
 }
