+++ src/main/java/com/bachelor/boostr/model/Price.java
@@ -3,16 +3,13 @@ package com.bachelor.boostr.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 public class Price {
 
-    private BigDecimal price;
+    private Double price;
 
     boolean isMonthlyOrEverydayPayments;
-
 
 }
